use `u32` as tag type in test cases

+++ serialization_helpers/tests/serialization_deserialization.rs
@@ -23,7 +23,7 @@ enum TestEnum {
 
 struct TestCase {
     data: TestEnum,
-    expected_tag: u16,
+    expected_tag: u32,
 }
 
 #[test]
